requirements.txt
Django
djangorestframework
psycopg2-binary==2.9.1
gunicorn
django-ratelimit
